Feat: 유저 이미지 생성 API 수정
